Address another low-risk Sonar cleanup batch for redundant numeric casts.
This batch removes redundant long/double casts in arithmetic and comparisons while preserving behavior under spotless formatting and -Pmin test validation.

Made-with: Cursor

+++ jacop-core/src/main/java/org/jacop/constraints/knapsack/KnapsackItem.java
@@ -92,7 +92,7 @@ public final class KnapsackItem implements Comparable<KnapsackItem> {
    */
   public int compareTo(KnapsackItem that) {
 
-    long comparison = (long) weight * (long) that.profit - (long) profit * (long) that.weight;
+    long comparison = (long) weight * that.profit - (long) profit * that.weight;
 
     if (comparison == 0) {
