+++ jacop-floats/src/main/java/org/jacop/examples/floats/Wilkinson.java
@@ -84,7 +84,7 @@ public class Wilkinson {
     FloatVar[] temp = new FloatVar[20];
     for (int i = 0; i < 20; i++) {
       temp[i] = new FloatVar(store, "temp[" + i + "]", minFloat, maxFloat);
-      FloatVar c = new FloatVar(store, (double) i + 1, (double) i + 1);
+      FloatVar c = new FloatVar(store, i + 1.0, i + 1.0);
       store.impose(new PplusQeqR(x, c, temp[i]));
     }
